+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [Fix:fix javadoc.](https://github.com/Tencent/spring-cloud-tencent/pull/721)
 - [Code optimization & correct comment](https://github.com/Tencent/spring-cloud-tencent/pull/722)
 - [fix:fix discovery junit.](https://github.com/Tencent/spring-cloud-tencent/pull/730)
+- [Test:add sct-stater-polaris-router juint.](https://github.com/Tencent/spring-cloud-tencent/pull/731)
 - [adapt polaris-java 1.10.1 version](https://github.com/Tencent/spring-cloud-tencent/pull/746)
 - [Optimize: change RouteArgument.buildCustom to RouteArgument.fromLabel](https://github.com/Tencent/spring-cloud-tencent/pull/749)
 - [Optimize: get service instances by Flux.blockLast() to resolve concurrent problem](https://github.com/Tencent/spring-cloud-tencent/pull/764)